add residual viewing

+++ analyzer/io/analyzermsgsender.cpp
@@ -13,6 +13,6 @@ void AnalyzerMsgSender::msgOut( const QString& strMsg , QtMsgType eMsgLevel)
 {    
     GitlEvent cEvt(g_strStatusMsgEvent);
     cEvt.setParameter("msg_detail", strMsg);
-    cEvt.setParameter("msg_level",eMsgLevel);
+    cEvt.setParameter("msg_level",(int)eMsgLevel);
     dispatchEvt(cEvt);
 }
